+++ backend/canisters/proposals_bot/impl/src/lib.rs
@@ -3,6 +3,7 @@ use crate::timer_job_types::TimerJob;
 use candid::{CandidType, Principal};
 use canister_state_macros::canister_state;
 use canister_timer_jobs::TimerJobs;
+use fire_and_forget_handler::FireAndForgetHandler;
 use serde::{Deserialize, Serialize};
 use sha2::{Digest, Sha256};
 use std::cell::RefCell;
@@ -79,6 +80,8 @@ struct Data {
     pub timer_jobs: TimerJobs<TimerJob>,
     #[serde(default)]
     pub registry_synced_up_to: TimestampMillis,
+    #[serde(default)]
+    pub fire_and_forget_handler: FireAndForgetHandler,
     pub test_mode: bool,
 }
 
@@ -91,9 +94,9 @@ impl Data {
         governance_principals: HashSet<Principal>,
         user_index_canister_id: CanisterId,
         group_index_canister_id: CanisterId,
+        registry_canister_id: CanisterId,
         cycles_dispenser_canister_id: CanisterId,
         nns_governance_canister_id: CanisterId,
-        registry_canister_id: CanisterId,
         test_mode: bool,
     ) -> Data {
         Data {
@@ -101,12 +104,13 @@ impl Data {
             governance_principals,
             user_index_canister_id,
             group_index_canister_id,
+            registry_canister_id,
             cycles_dispenser_canister_id,
             nns_governance_canister_id,
-            registry_canister_id,
             finished_proposals_to_process: VecDeque::new(),
             timer_jobs: TimerJobs::default(),
             registry_synced_up_to: 0,
+            fire_and_forget_handler: FireAndForgetHandler::default(),
             test_mode,
         }
     }