+++ backend/canisters/registry/c2c_client/src/lib.rs
@@ -5,3 +5,4 @@ use registry_canister::*;
 generate_c2c_call!(c2c_nervous_systems);
 
 // Updates
+generate_c2c_call!(c2c_set_submitting_proposals_enabled);